modules/locationsapi_post/requirements.txt
protobuf==3.18.1
grpcio==1.41.1
grpcio-tools==1.41.0
kafka-python==2.0.2